+++ src/main/kotlin/org/wfanet/measurement/duchy/daemon/mill/liquidlegionsv2/ReachOnlyLiquidLegionsV2Mill.kt
@@ -121,7 +121,7 @@ class ReachOnlyLiquidLegionsV2Mill(
   clock: Clock = Clock.systemUTC(),
   private val parallelism: Int = 1,
 ) :
-  LiquidLegionsV2MillBase(
+  LiquidLegionsV2Mill(
     millId,
     duchyId,
     signingKey,
