+++ src/test/kotlin/org/wfanet/measurement/duchy/daemon/mill/liquidlegionsv2/BUILD.bazel
@@ -1,12 +1,12 @@
 load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_test")
 
 kt_jvm_test(
-    name = "LiquidLegionsV2MillTest",
-    srcs = ["LiquidLegionsV2MillTest.kt"],
-    test_class = "org.wfanet.measurement.duchy.daemon.mill.liquidlegionsv2.crypto.LiquidLegionsV2MillTest",
+    name = "ReachFrequencyLiquidLegionsV2MillTest",
+    srcs = ["ReachFrequencyLiquidLegionsV2MillTest.kt"],
+    test_class = "org.wfanet.measurement.duchy.daemon.mill.liquidlegionsv2.crypto.ReachFrequencyLiquidLegionsV2MillTest",
     deps = [
         "//imports/java/io/opentelemetry/api",
-        "//src/main/kotlin/org/wfanet/measurement/duchy/daemon/mill/liquidlegionsv2:liquid_legions_v2_mill",
+        "//src/main/kotlin/org/wfanet/measurement/duchy/daemon/mill/liquidlegionsv2:reach_frequency_liquid_legions_v2_mill",
         "//src/main/kotlin/org/wfanet/measurement/duchy/daemon/testing",
         "//src/main/kotlin/org/wfanet/measurement/duchy/db/computation",
         "//src/main/kotlin/org/wfanet/measurement/duchy/db/computation/testing",
